refactored Radio.py--split each radio to own file
